remove decimal kernels, they're in arrow-rs now
