+++ datafusion/common/Cargo.toml
@@ -40,10 +40,10 @@ pyarrow = ["pyo3", "arrow/pyarrow"]
 
 [dependencies]
 apache-avro = { version = "0.14", default-features = false, features = ["snappy"], optional = true }
-arrow = { version = "23.0.0", default-features = false }
+arrow = { path = "../../../arrow-rs/arrow", default-features = false }
 cranelift-module = { version = "0.87.0", optional = true }
 object_store = { version = "0.5.0", default-features = false, optional = true }
 ordered-float = "3.0"
-parquet = { version = "23.0.0", default-features = false, optional = true }
+parquet = { path = "../../../arrow-rs/parquet", default-features = false, optional = true }
 pyo3 = { version = "0.17.1", optional = true }
 sqlparser = "0.23"
