+++ datafusion/core/fuzz-utils/Cargo.toml
@@ -23,6 +23,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-arrow = { version = "23.0.0", features = ["prettyprint"] }
+arrow = { path = "../../../../arrow-rs/arrow", features = ["prettyprint"] }
 env_logger = "0.9.0"
 rand = "0.8"
