+++ datafusion/core/tests/sql/decimal.rs
@@ -582,20 +582,20 @@ async fn decimal_arithmetic_op() -> Result<()> {
         "+---------------------------------------+",
         "| decimal_simple.c1 / decimal_simple.c5 |",
         "+---------------------------------------+",
-        "| 0.7142857142857143296                 |",
+        "| 0.7142857142857142857                 |",
         "| 0.8000000000000000000                 |",
-        "| 1.0526315789473683456                 |",
+        "| 1.0526315789473684210                 |",
         "| 0.9375000000000000000                 |",
-        "| 0.8571428571428571136                 |",
-        "| 2.7272727272727269376                 |",
-        "| 0.9090909090909090816                 |",
+        "| 0.8571428571428571428                 |",
+        "| 2.7272727272727272727                 |",
+        "| 0.9090909090909090909                 |",
         "| 1.0000000000000000000                 |",
         "| 1.0000000000000000000                 |",
-        "| 0.9090909090909090816                 |",
-        "| 0.9615384615384614912                 |",
-        "| 0.6410256410256410624                 |",
-        "| 1.5151515151515152384                 |",
-        "| 0.7352941176470588416                 |",
+        "| 0.9090909090909090909                 |",
+        "| 0.9615384615384615384                 |",
+        "| 0.6410256410256410256                 |",
+        "| 1.5151515151515151515                 |",
+        "| 0.7352941176470588235                 |",
         "| 0.5000000000000000000                 |",
         "+---------------------------------------+",
     ];
