+++ datafusion/expr/Cargo.toml
@@ -36,6 +36,6 @@ path = "src/lib.rs"
 
 [dependencies]
 ahash = { version = "0.8", default-features = false, features = ["runtime-rng"] }
-arrow = { version = "23.0.0", default-features = false }
+arrow = { path = "../../../arrow-rs/arrow", default-features = false }
 datafusion-common = { path = "../common", version = "12.0.0" }
 sqlparser = "0.23"
